% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binscatteR.R
\name{binscatter}
\alias{binscatter}
\title{Makes a bin scatter plot}
\usage{
binscatter(data = NULL, x = NULL, y = NULL, weights = NULL,
  numbins = 20, xlab = NULL, ylab = NULL, title = NULL,
  se = FALSE, smooth = FALSE)
}
\arguments{
\item{data}{data}

\item{x}{the x variable}

\item{y}{the y variable}

\item{weights}{optional weights}

\item{numbins}{number of bins. defaults to 20}

\item{xlab}{x label for plot. defaults to name of x}

\item{ylab}{y label for plot. defaults to name of y}

\item{title}{title for plot. defaults to binned scatter of y on x}

\item{se}{logical for whether to plot standard error of mean estimate. defaults to false}

\item{smooth}{logical for whether to plot smoother. defaults to false}
}
\value{
binned_df (the binned data) and plot_out (the plot)
}
\description{
Makes a bin scatter plot
}
