% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMeanVal.R
\name{GetMeanVal}
\alias{GetMeanVal}
\title{Get mean of \code{var} in an interval}
\usage{
GetMeanVal(interval, vals)
}
\arguments{
\item{interval}{Interval to get the mean of \code{var} on (character).}

\item{vals}{Values of \code{var}.}
}
\description{
Get mean of \code{var} in an interval
}
