% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMidpoint.R
\name{GetMidpoint}
\alias{GetMidpoint}
\title{Get midpoint of an interval}
\usage{
GetMidpoint(interval)
}
\arguments{
\item{interval}{Interval to get the midpoint of (character).}
}
\description{
Get midpoint of an interval
}
