% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeBinIndicator.R
\name{MakeBinIndicator}
\alias{MakeBinIndicator}
\title{Create bin indicators and bin centers for \code{x_var}}
\usage{
MakeBinIndicator(
  var,
  data,
  nBins,
  binType,
  intercept,
  min_obs,
  is_xvar = FALSE
)
}
\arguments{
\item{var}{Var}

\item{data}{Data}

\item{nBins}{Number of bins}

\item{binType}{Type of bin}

\item{intercept}{Boolean for intercept}

\item{min_obs}{Minimum number of observations per bin}

\item{is_xvar}{Boolean for whether \code{var} is \code{x_var}}
}
\description{
Create bin indicators and bin centers for \code{x_var}
}
