% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinScatter.R
\name{BinScatter}
\alias{BinScatter}
\title{Make data for a binned scatter plot and produce it.}
\usage{
BinScatter(
  data,
  y_var,
  x_var,
  linpartial_var = NULL,
  binpartial_var = NULL,
  n_bins = 20,
  bin_type = "quantile",
  n_partial_bins = NULL,
  partial_bin_type = NULL,
  min_obs = 10,
  min_partial_obs = NULL,
  intercept = FALSE,
  ci = NULL,
  drop_na = TRUE,
  recenter = TRUE,
  tab_path = NULL,
  plot_path = NULL,
  plot_xlab = NULL,
  plot_ylab = NULL,
  plot_xlim = NULL,
  plot_ylim = NULL,
  point_color = "black",
  point_fill = NA,
  point_shape = 19,
  point_size = 1.5,
  bar_width = 0.5,
  bar_size = 0.5,
  fig_width = 7,
  fig_height = 7,
  axis_title_x_size = NULL,
  axis_title_y_size = NULL,
  axis_text_size = NULL
)
}
\arguments{
\item{data}{Data frame for binned scatter plot.}

\item{y_var}{Name of LHS variable to appear on the y-axis (string).}

\item{x_var}{Name of RHS variable to appear on the x-axis (string).}

\item{linpartial_var}{Name(s) of RHS variable(s) to partial \code{x_var} on linearly (NULL, string, or
list of strings).
\itemize{
\item Note: categorical variables need to be one-hot coded, and are otherwise assumed continuous.
}}

\item{binpartial_var}{Name(s) of RHS variable(s) to partial \code{x_var} on with bins (NULL, string, or
list of strings).}

\item{n_bins}{Number of bins for \code{x_var} (integer).}

\item{bin_type}{Method to create bins for \code{x_var} ("uniform" or "quantile")
\itemize{
\item "uniform" bins are equally spaced over the support of \code{x_var}.
\item "quantile" bins have approximately the same number of observations in each bin.
}}

\item{n_partial_bins}{Number of bins for variables in \code{binpartial_var} (integer, default is \code{n_bins}).}

\item{partial_bin_type}{Method to create bins for variables in \code{binpartial_var} ("uniform" or
"quantile", default is \code{bin_type}).}

\item{min_obs}{Minimum number of observations per bin for \code{x_var} (integer, default is 10).}

\item{min_partial_obs}{Minimum number of observations per bin for variables in \code{binpartial_var}
(integer, default is \code{min_obs}).}

\item{intercept}{Whether an intercept is included in the regression (boolean).}

\item{ci}{Whether to include confidence intervals and confidence level (NULL or numeric in (0,1)).
\itemize{
\item If not NULL, std. err. of the partialed means of \code{y_var} will be returned in the output dataframe.
\item If specified but non-numeric or out of range, the default confidence level is 0.95.
}}

\item{drop_na}{Whether to drop observations with NA in \code{x_var}, \code{y_var}, \code{linpartial_var}, and
\code{binpartial_var} (boolean).
\itemize{
\item If \code{TRUE}, \code{NA} will be dropped automatically with a warning message.
\item If \code{FALSE}, returns an error if \code{NA} is detected.
}}

\item{recenter}{Whether to recenter partialed means of \code{y_var} (boolean).}

\item{tab_path}{Path to output the regression table (string).}

\item{plot_path}{Path to save the binned scatter plot (string).}

\item{plot_xlab}{X-axis title for the binned scatter plot (string, default is \code{x_var}).}

\item{plot_ylab}{Y-axis title for the binned scatter plot (string, default is \code{y_var}).}

\item{plot_xlim}{Range of x-axis for the binned scatter plot (\code{c(min, max)}, default is
ggplot default).}

\item{plot_ylim}{Range of y-axis for the binned scatter plot (\code{c(min, max)}, default is
ggplot default).}

\item{point_color}{Color aesthetic to be used with \code{geom_point()}.}

\item{point_fill}{Fill aesthetic to be used with \code{geom_point()}.}

\item{point_shape}{Shape aesthetic to be used with \code{geom_point()}.}

\item{point_size}{Size aesthetic to be used with \code{geom_point()}.}

\item{bar_width}{Width aesthetic to be used with \code{geom_errorbar()}.}

\item{bar_size}{Size aesthetic to be used with \code{geom_errorbar()}.}

\item{fig_width}{Plot width in inches (numeric).}

\item{fig_height}{Plot height in inches (numeric).}

\item{axis_title_x_size}{Text size for x-axis labels.}

\item{axis_title_y_size}{Text size for y-axis labels.}

\item{axis_text_size}{Text size for tick labels along axes.}
}
\description{
Make data for a binned scatter plot and produce it.
}
