% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMeanVal.R
\name{GetMeanVal}
\alias{GetMeanVal}
\title{Get mean of \code{x_var} in an interval}
\usage{
GetMeanVal(interval, vals)
}
\arguments{
\item{interval}{Interval}

\item{vals}{Values}
}
\description{
Get mean of \code{x_var} in an interval
}
